Beginning of the Frontend work
I started a basic web responsive layout for our app. I set up the base for the Routes we will use or the Routes that will be added.  I created the login/register interface where the form and UI changes depending on wether the user wants to register or login.