Created the "My Offer" page
Created the frontend of My Offer page. We search the through the offers from the database and we filter them by OfferCreator and only keep the ones that match with the current user. Then we take the Property ID from those offers and fetch the Properties from the data base that matches with the Property IDs we have. Then we display them. With the help of Peno we set up the backend get requests for the offers.