Adding more features
I added the ability to Accept or Reject an offer in the OfferManagement page. Additionnaly, i added a button to undo the latter. Also in the offer page I added a delete button in case the user would like to withdraw a button.  I also added another function that help us access the login token, and made sure it's been deleted.